More cabal file cleanups
This organizes some sections, and shuffles around a few fields here and there
(for example, moving 'default-languages' to the top of the stanza or close to
it, since it's common amongst all of them and easier to see differences later
on). Includes some extra whitespace to spice things up a bit.

Signed-off-by: Austin Seipp <aseipp@pobox.com>