Fix #13 - make usage of ForeignPtrs safe
Previously, we were taking the `Ptr` values out of a `ForeignPtr`
to offset/peek into them. Unfortunately if the `ForeignPtr`s last
reference was where it was deconstructed, the GC might kick in
and collect it, by the time we have the `Ptr` handed to us.

Instead, make sure we use `withForeignPtr`, which properly uses
the `touch#` primop to ensure the original `ForeignPtr` stays
alive for the duration of the time we use the `Ptr`.

Signed-off-by: Austin Seipp <austin@well-typed.com>