FlatTerm: fix some edge cases in decoding Words
When we convert an `Encoding` to a FlatTerm, and ask to encode a
`Word` value, we check if it's in the range of `maxBound :: Int`
first, and if it is too large, we store it in an `Integer`. But,
Word always has a larger maxBound than Int.

The FlatTerm decoder wasn't taking this case into account; so
if you have `(maxBound :: Int) + 1` as a Word (which is perfectly
OK), it would get stored inside the FlatTerm using an Integer,
and the decoder only took the Int case into account.

Signed-off-by: Austin Seipp <austin@well-typed.com>