bench: Add vector instance benchmarks
These are the benchmarks from #42, reworked to give the benchmarks
slightly better names. The prior ones had a name like
'500/encode/boxed', '5000/encode/boxed', and so on, meaning it was
impossible to distinguish them by a prefix match, like if I wanted
criterion to only run the boxed encoder tests.

The new benchmarks have a form like:

   instance/vector/boxed/encode/500
   instance/vector/boxed/encode/5000
   instance/vector/boxed/encode/50000
   instance/vector/boxed/encode/500000
   instance/vector/boxed/decode/500
   instance/vector/boxed/decode/5000
   instance/vector/boxed/decode/50000
   instance/vector/boxed/decode/500000

Which makes it much easier to look exactly at what you're improving
by simply providing criterion the prefix of the benchmark name.

Closes #42.

Signed-off-by: Austin Seipp <austin@well-typed.com>