Read: implement 32bit support
Turned out to be much easier than I thought, and all the test seem to
be clean on my 32bit Linux box with a quick torture test of thousands
of QC rounds at low GC settings, so we'll cautiously push this for
now. And hopefully clean up the CI, too.

Fixes #20.

Signed-off-by: Austin Seipp <austin@well-typed.com>