Fix warnings
There's been a couple of unnecessary imports in the main package and a
bunch of various warnings in the benchmarking code. This fixes almost
all of them (except for bench/versus/Macro/CBOR where I decided to simply
disable the warning about missing signatures - it doesn't seem to be
that valuable in this case and would require a fair amounut of
boilerplate).

Closes #57.

Signed-off-by: Michal Terepeta <michal.terepeta@gmail.com>
Signed-off-by: Austin Seipp <austin@well-typed.com>