ByteOrder: do proper narrowing in grabWord{16,32,64}
The grabWord{16,32,64} utilities in ByteOrder use underlying byteSwap
primitives if necessary to read big-endian integers quickly. However,
the underlying primitives do not specify any narrowing of range when
doing the conversion.

This results in some CMM to the effect of:

    (_c3rw::I64) = call MO_BSwap W16(%MO_UU_Conv_W16_W64(I16[c3rr_str]));
    I64[Sp - 24] = c3rD;
    R4 = GHC.Types.[]_closure+1;
    R3 = _c3rw::I64;

The store to R3 is the important part. This generates some LLVM code
like:

    %ln3sm = call ccc i16 (i16)* @llvm.bswap.i16( i16 %ln3sl )
    %ln3sn = sext i16 %ln3sm to i64
    store i64 %ln3sn, i64* %lc3rJ
    ...
    %ln3st = load i64* %lc3rJ
    store i64 %ln3st, i64* %R3_Var

However, when LLVM's `sext` instruction is called, it works by merely
copying the sign bit until it extends the value appropriately,
e.g. the instr `sext i8 -1 to i16 == 65535`. But the results of bswap
on the higher parts of a word aren't really defined, so it's not clear
what state the value will be in when `sext` happens. When reading an
Int type, LLVM will merely propogate the sign bit to extend the word,
and stores that. When the result is read back, the sign bit will be
set (and possibly other bits), indicating overflow (and thus giving
back a negative result).

However, after properly narrowing using the narrow* primitives:

    (_c3rk::I64) = call MO_BSwap W16(%MO_UU_Conv_W16_W64(I16[c3rf_str]));
    I64[Sp - 24] = c3ru;
    R4 = GHC.Types.[]_closure+1;
    R3 = %MO_UU_Conv_W16_W64(%MO_UU_Conv_W64_W16(_c3rk::I64));

We get a proper `trunc`/`zext` after the load:

    %ln3sw = call ccc i16 (i16)* @llvm.bswap.i16( i16 %ln3sv )
    %ln3sx = sext i16 %ln3sw to i64
    store i64 %ln3sx, i64* %lc3rQ
    ...
    %ln3sD = load i64* %lc3rQ
    %ln3sE = trunc i64 %ln3sD to i16
    %ln3sF = zext i16 %ln3sE to i64
    store i64 %ln3sF, i64* %R3_Var

With that, the resulting sign bits from `sext` are removed with a
combination of `trunc`/`zext`, so the correct result is given.

This patch properly narrows grabWord{16,32,64} in the same style the
GHC base libraries do (see the source code to GHC.Word).

With this fix, the entire testsuite fully passes with the setup of GHC
7.10.3/LLVM 3.5.2 and GHC 8.0.1/LLVM 3.7.1. However, the performance
impact of LLVM seems to be marginal (or even worse) overall in the
benchmarks.

See also https://ghc.haskell.org/trac/ghc/ticket/12095

Fixes #67.

Signed-off-by: Austin Seipp <austin@well-typed.com>