Rename PeekLength and make it use Int# rather than Int
Rename PeekLength to PeekAvailable, and peekLength/peekAvailable.
Change the representation in DecodeAction to pass an unboxed Int#
rather than a boxed Int, like we do in all the other cases.

Also handle it properly at the top level. It does not require input so
in principle it's wrong to require input first, same as Done or Fail.