demos: Refactor demo exes into a more useful 'cbor-tool' (#84)
This removes most of the old demo code and refactors it into a single, more
general tool that fixes a few issues from the previous ones. The demo code
for doing Aeson conversions is still there, but only the boilerplate, so you
don't have to look through the rest of it.

Fixes #76. Closes #75. Closes #77.

Signed-off-by: Austin Seipp <austin@well-typed.com>