tool: add --lenient flag for JSON decoding
When decoding arbitrary CBOR values to JSON, the RFC specifies that the
only legitimate key type in a CBOR map for a JSON map is a UTF8 string.

However, general Haskell data types, or any arbitrary CBOR, likely has
integers for map keys, meaning they cannot be easily turned into JSON.
This is the right behavior: CBOR maps are more expressive, allowing more
than just strings as keys. So naively pretty-printing the key 1 and key
"1" both as a string will cause collisions in the output, giving you
invalid JSON.

However, JSON dumps are still way more human readable as a diagnostic
output. Thus, a new --lenient flag allows integers and booleans to pass
as valid keys for a Map (they're simply converted to Text).

This also Value decoder to be more general and robust, as a result.

Signed-off-by: Austin Seipp <austin@well-typed.com>