Magic: do not export withBsPtr
Aside from no longer being needed anywhere else, this is easily one of the more
dangerous primitives to misuse, because it does not prevent you from writing to
the (immutable) Ptr, among other things.

If anyone reasonably might have a use for this in the future I'm open to
re-exporting it, but I can't imagine any at the moment.

Signed-off-by: Austin Seipp <austin@well-typed.com>