Add a 'nix-shell' script to quickly launch a dev environment
This is a bit less flexible, but easier than writing a big shell.nix to tie all
the packages together. That could be done in the future, but for now, this
works well for just doing rapid 'cabal new-build' development. It passes any
given arguments directly to 'nix-shell'. For example:

    $ ./nix-shell --run \
        'cabal new-build all --enable-tests --enable-benchmarks'

Signed-off-by: Austin Seipp <aseipp@pobox.com>