Clean up README
The READMEs were misleading in various ways due to me predicating a lot of
future stuff didn't happen in a prompt manner. So, remove all that junk, and
remove the old broken Stackage/AppVeyor badges, for now.

Once we bring back AppVeyor and get a Stackage LTS release, we can add those
back.

This also removes the serialise/README.md file, since it was mostly redundant
in favor of the top level one.

Closes #165.

Signed-off-by: Austin Seipp <aseipp@pobox.com>