Add peekByteOffset primitive
Get the current ByteOffset in the input byte sequence of the Decoder.
The Decoder does not provide any facility to get at the input data
directly (since that is tricky with an incremental decoder). The next
best is this primitive which can be used to keep track of the offset
within the input bytes that makes up the encoded form of a term.

By keeping track of the byte offsets before and after decoding a
subterm (a pattern captured by decodeWithByteSpan) and if the overall
input data is retained then this is enables later retrieving the span
of bytes for the subterm.

Also cover the PeekByteOffset case in fromFlatTerm. It doesn't really
make sense here. Use 0.