Add tests for peekByteOffset
This should be pretty comprehensive. The main property we test is that
if we decode a Term where we track the byte span of the whole term and
all sub-terms then the bytes we get do indeed deserialise to the same
sub-term, or reserialising the term gives us the bytes.

Overall this tells us that we must be getting the byte offsets right.

Also cover non-canonical case for peekByteOffset tests. Demonstrate
that for non-canonical encodings, re-serialising the term does not give
the captured byte span. This is of course on of the main points of
using peekByteOffset in the first place: to get at the raw data in the
cases where it cannot be recovered by re-serialising.