Adjust eqTerm to distinguish different NaNs
And adjust uses of eqTerm to use canonicaliseNaNs as appopriate.

Previously, the eqTerm would compare equality but define all NaN values
to be equivalent. This was needed in some properties because some
properties do a round trip from Term to RefImpl.Term, or a round trip of
serialising and deserialising, either of which will replace NaNs by the
canonical NaN representation. So when comparing the result of the round
trip with the expected value, we need to canonicalise the NaNs first.

This patch makes that process more explicit, by making eqTerm be exact
when it comes to the various NaNs and adding and using canonicaliseNaNs
in those properties where it is needed.