Add new generalised test properties
Now with a commuting diagram to explain why these are the right
properties.

They are generalised compared to the existing Term-only properties,
which will allow us to cover individual token types, not just Terms.

This partially obsoletes some of the Term properties, but this patch
does not remove any yet.