Add an encodePreEncoded function
This allows pre-encoded CBOR data to be included into an Encoding. This
is useful in cases where one has known-valid encoded CBOR data, e.g. on
disk, that you want to include into a larger CBOR data stream. This
makes it possible in such cases to avoid decoding and re-encoding.