Fetch remote cborg-json benchmark data
Since we do not include the benchmark data into the tarball then by
default the benchmark does not even compile (since we used file-embed).

This patch switches to fetching it remotely at the beginning of the
benchmark, if it's not already present, using the Criterion.env feature.

Hopefully this will also fix it on CI.