Add round-trip testing for to/fromFlatTerm
This is very closely based on patches from Deepak Kapiswe, and thanks to
him for doing all the hard work here.

I have simplified things slightly from the original proposed property.
The property that Deepak proposed was indeed true but it was more
complicated than it needed to be because the code didn't actually
satisfy a simpler property, but it should have done!

The simpler property clealy ought to be true. So this helped clarify my
thinking on what FlatTerm is and what toFlatTerm is for. I originally
thought that it should reflect 1:1 what an Encoding looks like.

The normal full serialisation to bytes satisfies a round trip property,
even though the full serialisation does a fair degree of
canonicalisation (e.g. numbers in their minimal form). The FlatTerm
conversions should satisfy an equivalent round trip property. In which
case, the question arises: where is the canonicalisation done? In
serialisation to bytes the canonicalisation happens as the 'Encoding'
tokens are written out to bytes. So by analogy, for FlatTerm the
canonicalisation should happen in the conversion from 'Encoding' to
'FlatTerm'. This also makes sense in that the flat term should show you
want the external tokens will be, not some intermediate form. And we
already do that to some degree, e.g. with toFlatTerm of a small Integer
will produce a TkInt, because that's how it will actually be encoded.

The only place we were not doing this kind of conversion was half-width
floats and float NaNs. The simplified test pointed this out.

So this patch adds the round trip test and fixes the representation of
the TkFloat16 to be a Half, and fixes toFlatTerm for NaNs.