Fix implementation of peekTokenType for over-long encodings
The peekTokenType distinguishes not just the major CBOR types but also
the ones encoded a "simple" types or tagged types, such as bool and big
integer.

The implementation of this only correctly recognised the normal 1-byte
encodings of these, and not the non-canonical over-long encodings. For
example, it's possible to use a 32bit encoding of the tag 2, rather than
the normal encoding that puts it into the header byte.