Add property test for comparing decoding on invalid inputs
We generate byte sequences corresponding to sequences of valid tokens
and then try to decode them as CBOR terms. This will of course often
fail since not all token sequences correspond to valid terms. We check
that both the reference decoder and the fast decoder agree that the
bytes are valid (and same value), or invalid.

This exposes inconsistencies between the reference and fast decoder.
They all turned out to be of the form where the fast decoder did not
handle over-long encodings of certain things: simple and tagged values
and thus things based on these: bools, null, big integers.

This test is inspired by a test contributed by Deepak Kapiswe. His
original contribution tested validateFlatTerm against the reference
decoder.