Don't use unlifted numerics in canonical Read implementation
This depended upon the fact that previously the lifted sized integral
types (e.g. `Int8`) were represented internally by the unsized unlifted
integral types (e.g. `Int#`). With GHC 9.2 this is no longer true.

Consequently, it is easier to just rely on the lifted types and rely
upon the compiler to worker-wrapper away the boxes. I have verified that
with GHC 8.10.7 the simplified Core produced is equivalent up to
alpha-renaming.