WIP: experimentally use dupIO to fix #318
This adds a new token type TkDup for use when encoding, that duplicates
the contained closure. This means that encodings can explicitly duplicate
thunks that may otherwise end up in the old generation.  It isn't obvious
to me whether this is enough or whether calling dupIO every time would be
better (and whether that might have an unreasonable performance cost).